reorganizing sample apps folder structure
