+++ tests/integration/common/apps/HelloWorldApp/components/Contact.js
@@ -1,4 +1,4 @@
-import  { parseJsonSchema } from '../../../../../src/utils/JsonSchemaToReact';
+import  { parseJsonSchema } from '../../../../../../src/utils/JsonSchemaToReact';
 
 import pageStyle from './pageStyles';
 
@@ -9,7 +9,7 @@ export default ({ match, appName }) => {
 		children: [
 			{
 				component: 'Text',
-				text: 'Welcome to the home page!'
+				text: 'Welcome to the contact page!'
 			}
 		]
 	};
