+++ tests/integration/common/apps/HelloWorldApp/components/pageStyles.js
