+++ tests/integration/common/apps/TodoMVCApp/actions/index.spec.js
