+++ tests/integration/common/apps/TodoMVCApp/components/Footer.js
