+++ tests/integration/common/apps/TodoMVCApp/components/Footer.spec.js
