+++ tests/integration/common/apps/TodoMVCApp/components/MainSection.js
