+++ tests/integration/common/apps/TodoMVCApp/components/TodoItem.spec.js
