+++ tests/integration/common/apps/TodoMVCApp/components/TodoTextInput.js
