+++ tests/integration/common/apps/TodoMVCApp/components/TodoTextInput.spec.js
