+++ tests/integration/common/apps/TodoMVCApp/constants/ActionTypes.js
