+++ tests/integration/common/apps/TodoMVCApp/containers/App.js
