+++ tests/integration/common/apps/TodoMVCApp/reducers/index.js
