+++ tests/integration/common/plugins/ApolloPlugin/defaultParams.js
